+++ packages/vendure-order-client/CHANGELOG.md
@@ -1,6 +1,6 @@
 # 2.8.1 (2023-12-01)
 
-- Update the getOrderByCode to be polling for 30 seconds
+- Update the getOrderByCode and add pollOrderByCode function
 
 # 2.8.0 (2024-01-04)